requirements.txt
boto3
pandas
numpy
seaborn
matplotlib
scikit-learn
transformers
torch
dill
tensorflow
openpyxl
zemberek-python
snscrape
eksipy
asyncio
-e .